updating Tests with review comments
