+++ subprojects/scala/src/main/java/org/gradle/api/tasks/scala/ScalaCompile.java
@@ -15,11 +15,8 @@
  */
 package org.gradle.api.tasks.scala;
 
-import org.gradle.api.Incubating;
 import org.gradle.api.InvalidUserDataException;
-import org.gradle.api.JavaVersion;
 import org.gradle.api.file.FileCollection;
-import org.gradle.api.file.FileTree;
 import org.gradle.api.internal.file.FileResolver;
 import org.gradle.api.internal.project.ProjectInternal;
 import org.gradle.api.internal.tasks.scala.ScalaCompileSpec;
@@ -27,10 +24,7 @@ import org.gradle.api.internal.tasks.scala.ScalaCompilerFactory;
 import org.gradle.api.internal.tasks.scala.ScalaJavaJointCompileSpec;
 import org.gradle.api.tasks.CacheableTask;
 import org.gradle.api.tasks.Classpath;
-import org.gradle.api.tasks.Input;
 import org.gradle.api.tasks.Nested;
-import org.gradle.api.tasks.PathSensitive;
-import org.gradle.api.tasks.PathSensitivity;
 import org.gradle.language.scala.tasks.AbstractScalaCompile;
 import org.gradle.process.internal.worker.child.WorkerDirectoryProvider;
 import org.gradle.workers.internal.WorkerDaemonFactory;
@@ -67,15 +61,6 @@ public class ScalaCompile extends AbstractScalaCompile {
         return scalaClasspath;
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    @PathSensitive(PathSensitivity.NAME_ONLY)
-    public FileTree getSource() {
-        return super.getSource();
-    }
-
     public void setScalaClasspath(FileCollection scalaClasspath) {
         this.scalaClasspath = scalaClasspath;
     }
@@ -90,19 +75,6 @@ public class ScalaCompile extends AbstractScalaCompile {
 
     public void setZincClasspath(FileCollection zincClasspath) {
         this.zincClasspath = zincClasspath;
-    }
-
-    /**
-     * The Java major version of the JVM the Scala compiler is running on.
-     *
-     * @since 4.1
-     */
-    @Incubating
-    @Input
-    // We track this as an input since the Scala compiler output may depend on it.
-    // TODO: This should be replaced by a property in the Scala toolchain as soon as we model these.
-    protected String getJvmVersion() {
-        return JavaVersion.current().getMajorVersion();
     }
 
     /**
